Make CSMA_LIMIT=-90dBm (less sensitive when sending)
Sometimes (especially 433Mhz versions) are too sensitive or there is more noise in the channel than usual. Touching the antenna or coiling it helps a lot but a more permanent fix is to adjust the sensitivity.
This addresses the issue when send sensitivity is too high and the radio will spend a lot of time reading RSSI to make sure channel is clear, resulting in dropped ACKs.