Explicitly re-enable interrupts in receiveDone
The comment stated that setMode would re-enable interrupts, but it doesn't look like the code does that. This is causing problems for me on a SAMD21-based custom board - it causes subsequent calls to delay() to hang forever. The Arduino time is implemented by an interrupt every millisecond to increment the time, so I think that this is the culprit.